[MISC] Move seqan3::views::rank_to to seqan3/alphabet/views/
